s1

+++ pybricks/pyBricks1/wheelcontroller.py
@@ -8,9 +8,9 @@ from shared import Shared, Speed
 
 
 class WheelController:
-    # I measured manually and the wheel diameter is 5.6cm and the axle distance is 11.7cm float(117)
-    # when double wheels wheel diameter is 5.6cm and the axle distance is 17.4cm
-    __wheel_diameter_in_mm = float(55)
+    # I measured manually and the wheel diameter is 5.6cm and the axle distance is 11.7cm float(117) for blue wheels
+    # 2nd black wheels, wheel diameter is 5.5cm and the axle distance is 18.5cm
+    __wheel_diameter_in_mm = float(80)
     __axle_track_in_mm = float(185)
 
     __left_motor = Motor(Port.E, Direction.COUNTERCLOCKWISE)
